user and goal hook to load more data
